Add Rooms
We now have Rooms which have Trackers.
I have added the following rest apis for both Rooms and Trackers:
Where `{type}` is either `room` or `tracker`.
- `GET /{type}/all`               - Get all entities of the given type
- `POST /{type}/add`              - Add an entity using json
- `PUT /{type}/update`            - Update an entity's properties given its id
- `DELETE /{type}/remove?id={id}` - Remove an entity given its id
- `GET /room/get?id={id}`         - Get a room by its id

We now use CancellationToken everywhere.
This allows us to cancel a database request and or http request.

The UI now shows rooms and you are allowed to delete it.
Creating and updating rooms still needs to be added, in addition to
adding / updating / removing trackers under the rooms.