Refactoring
Seperate out IRoomRepository and ITrackerRepository

We are no longer using the database objects for serialization and
deserialization on the REST Api.
Now we have specific objects that represent the requests.
This means we no longer have data inside that are no longer used and we
no longer have to juggle entity framework and json serialization.

We now also have response objects.
These give the user of the api more information about the status of the
request.
For example adding a room now gives the user the id of the newly added
room.
Another example is that adding a tracker to a room is now fallible.
If the room id the user tries to add a tracker to does not exist we
return that the request was unsuccessful.