Add Charts for Rooms
Fixed getting AggregatedMeasurements
Added a page that takes a room id and displays graphs for Wifi and
Bluetooth devices.

TODO: Add a graph that shows the calculated aggregated person count.
TODO: Add a a date range control where you can get data for a specific
range and default it to a reasonable value.